Renamed package lex to scanner and moved some stuff around.
