+++ src/io/github/sboyanovich/scannergenerator/automata/NFA.java
@@ -1,6 +1,6 @@
 package io.github.sboyanovich.scannergenerator.automata;
 
-import io.github.sboyanovich.scannergenerator.lex.StateTag;
+import io.github.sboyanovich.scannergenerator.scanner.StateTag;
 
 import java.util.*;
 import java.util.function.Function;
