+++ src/io/github/sboyanovich/scannergenerator/scanner/NameDictionary.java
@@ -1,4 +1,4 @@
-package io.github.sboyanovich.scannergenerator.lex;
+package io.github.sboyanovich.scannergenerator.scanner;
 
 import java.util.ArrayList;
 import java.util.HashMap;
