+++ src/io/github/sboyanovich/scannergenerator/scanner/Position.java
@@ -1,4 +1,4 @@
-package io.github.sboyanovich.scannergenerator;
+package io.github.sboyanovich.scannergenerator.scanner;
 
 public class Position implements Comparable<Position> {
     private int line;
