+++ src/io/github/sboyanovich/scannergenerator/scanner/token/BasicToken.java
@@ -1,6 +1,6 @@
-package io.github.sboyanovich.scannergenerator.token;
+package io.github.sboyanovich.scannergenerator.scanner.token;
 
-import io.github.sboyanovich.scannergenerator.Fragment;
+import io.github.sboyanovich.scannergenerator.scanner.Fragment;
 
 /** Most basic, attributeless token. Useful for singleton lexical domain like keywords, op signs, parens etc.*/
 public class BasicToken extends Token {
