+++ src/io/github/sboyanovich/scannergenerator/scanner/token/DomainError.java
@@ -1,7 +1,7 @@
-package io.github.sboyanovich.scannergenerator.token;
+package io.github.sboyanovich.scannergenerator.scanner.token;
 
-import io.github.sboyanovich.scannergenerator.Fragment;
-import io.github.sboyanovich.scannergenerator.lex.Text;
+import io.github.sboyanovich.scannergenerator.scanner.Fragment;
+import io.github.sboyanovich.scannergenerator.scanner.Text;
 import io.github.sboyanovich.scannergenerator.utility.Utility;
 
 public enum DomainError implements DomainWithAttribute<String> {
