+++ src/io/github/sboyanovich/scannergenerator/tests/data/domains/SimpleDomains.java
@@ -1,10 +1,10 @@
 package io.github.sboyanovich.scannergenerator.tests.data.domains;
 
-import io.github.sboyanovich.scannergenerator.Fragment;
-import io.github.sboyanovich.scannergenerator.lex.Text;
-import io.github.sboyanovich.scannergenerator.token.BasicToken;
-import io.github.sboyanovich.scannergenerator.token.Domain;
-import io.github.sboyanovich.scannergenerator.token.Token;
+import io.github.sboyanovich.scannergenerator.scanner.Fragment;
+import io.github.sboyanovich.scannergenerator.scanner.Text;
+import io.github.sboyanovich.scannergenerator.scanner.token.BasicToken;
+import io.github.sboyanovich.scannergenerator.scanner.token.Domain;
+import io.github.sboyanovich.scannergenerator.scanner.token.Token;
 
 public enum SimpleDomains implements Domain {
     WHITESPACE {
@@ -89,6 +89,18 @@ public enum SimpleDomains implements Domain {
         @Override
         public Token createToken(Text text, Fragment fragment) {
             return new BasicToken(fragment, ESCAPED_RPAREN);
+        }
+    },
+    OPERATION {
+        @Override
+        public Token createToken(Text text, Fragment fragment) {
+            return new BasicToken(fragment, OPERATION);
+        }
+    },
+    KEYWORD {
+        @Override
+        public Token createToken(Text text, Fragment fragment) {
+            return new BasicToken(fragment, KEYWORD);
         }
     }
 }
