+++ src/io/github/sboyanovich/scannergenerator/tests/data/tokens/TNonTerminal.java
@@ -1,8 +1,8 @@
 package io.github.sboyanovich.scannergenerator.tests.data.tokens;
 
-import io.github.sboyanovich.scannergenerator.Fragment;
+import io.github.sboyanovich.scannergenerator.scanner.Fragment;
 import io.github.sboyanovich.scannergenerator.tests.data.domains.DomainsWithStringAttribute;
-import io.github.sboyanovich.scannergenerator.token.TokenWithAttribute;
+import io.github.sboyanovich.scannergenerator.scanner.token.TokenWithAttribute;
 
 public class TNonTerminal extends TokenWithAttribute<String> {
     public TNonTerminal(Fragment coords, String attribute) {
